functional delete contact but faulty passing of idc value

+++ app/src/main/java/com/mobdeve/s16/desembrana/annapatricia/PressSOS/Contact.java
@@ -2,7 +2,7 @@ package com.mobdeve.s16.desembrana.annapatricia.PressSOS;
 
 public class Contact {
 
-    private long id;
+    private int id;
     private String name;
     private String contactNumber;
 
@@ -11,7 +11,7 @@ public class Contact {
         this.contactNumber = contactNumber;
     }
 
-    public Contact (String name, String contactNumber, long id) {
+    public Contact (String name, String contactNumber, int id) {
         this.name = name;
         this.contactNumber = contactNumber;
         this.id = id;
@@ -19,7 +19,7 @@ public class Contact {
 
     public String getName() { return name; }
     public String getContactNumber() { return contactNumber; }
-    public long getId() {
+    public int getId() {
         return id;
     }
